Remove custom transform setting from jest configuration
