+++ src/setupTests.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-console */
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import '@testing-library/jest-dom/extend-expect';
+import '@testing-library/jest-dom';
 import 'jest-localstorage-mock';
 import './tests/initI18n';
 import { toHaveNoViolations } from 'jest-axe';
